refactor: 公开 Message 编码与执行生命周期原子
