+++ plugin_boundary_baseline.toml
@@ -52,17 +52,11 @@ R2 = [
     "plugins/codex/driver.py|core.net.http",
     "plugins/codex/responses.py|core.net.http",
     "plugins/content/api.py|agent.turn_effects",
-    "plugins/content/api.py|session.artifacts",
-    "plugins/delivery_policy/plugin.py|agent.restart",
     "plugins/markdown_memory/message_plugin.py|agent.llm_json",
     "plugins/markdown_memory/message_plugin.py|infra.persistence.json_store",
     "plugins/markdown_memory/store.py|infra.persistence.json_store",
     "plugins/message_push/plugin.py|agent.control.frame_book",
-    "plugins/message_push/plugin.py|agent.restart",
     "plugins/message_push/restart.py|agent.control.frame_book",
-    "plugins/message_push/restart.py|agent.restart",
-    "plugins/message_push/tool.py|session.artifacts",
-    "plugins/message_push/tool.py|session.message_codec",
     "plugins/models/content.py|agent.media",
     "plugins/models/litellm_catalog.py|agent.model_runtime.catalog.litellm_registry",
     "plugins/models/state.py|agent.plugins.snapshot",
@@ -75,9 +69,6 @@ R2 = [
     "plugins/prompt/text.py|agent.persona",
     "plugins/prompt/text.py|prompts.completion",
     "plugins/qq_sender/sender.py|session.artifacts",
-    "plugins/react/plugin.py|agent.plugins.snapshot",
-    "plugins/reply/completion.py|agent.restart",
-    "plugins/reply/follow.py|agent.restart",
     "plugins/scheduler/migration.py|infra.persistence.json_store",
     "plugins/scheduler/store.py|core.common.timekit",
     "plugins/scheduler/store.py|infra.persistence.json_store",
@@ -108,7 +99,6 @@ R2 = [
     "plugins/telegram_sender/sender.py|infra.channels.telegram_utils",
     "plugins/telegram_sender/sender.py|session.artifacts",
     "plugins/wake/dashboard.py|agent.plugins.snapshot",
-    "plugins/wake/runtime.py|agent.control.timer",
     "plugins/workbench_ui/dashboard.py|agent.plugins.snapshot",
     "plugins/workbench_ui/dashboard.py|infra.channels.message_view",
 ]
