+++ plugins/content/api.py
@@ -9,7 +9,7 @@ from types import MappingProxyType
 from typing import cast
 
 from agent.turn_effects import PostCommitEffect, post_commit_effect
-from session.artifacts import check_artifact_id
+from agent.plugin_composition.artifacts import check_artifact_id
 from agent.plugin_contracts import ContentPart, ContentReferences, Control, Input, Message
