+++ plugins/reply/completion.py
@@ -3,7 +3,7 @@ from contextlib import AbstractAsyncContextManager, AbstractContextManager
 from typing import Protocol
 
 from agent.plugin_composition import ServiceKey
-from agent.restart import ExternalRootPermit
+from agent.plugin_composition.tasks import ExternalRootPermit
 from agent.plugin_composition.messages import MessageReader
