+++ session/message_codec.py
@@ -1,106 +1,5 @@
-from __future__ import annotations
-
-import json
-from collections.abc import Mapping
-from typing import Literal, cast
-
-from session.message import (
-    Body,
-    CallRef,
-    ContentPart,
-    Control,
-    Input,
-    Output,
-    Part,
-    ToolCall,
-    ToolResult,
+"""旧 Core 路径兼容；Message 编解码实现由公开值模块拥有。"""
+from agent.plugin_contracts.message import (
+    body_to_dict as body_to_dict, json_value as json_value,
+    encode_body as encode_body, decode_body as decode_body,
 )
-
-
-from agent.plugin_contracts.message import body_to_dict as body_to_dict, json_value as json_value
-
-
-def encode_body(body: Body, *, allow_legacy: bool = True) -> str:
-    """保留已读消息的历史编码；新写入可显式拒绝旧表示。"""
-    data = body_to_dict(body)
-    if isinstance(body, ToolResult) and body._legacy_unknown:
-        if not allow_legacy:
-            raise ValueError("旧 unknown 工具结果只能重放已有消息，不能作为新消息写入")
-        data["outcome"] = "unknown"
-    return json.dumps(
-        data, ensure_ascii=False, sort_keys=True, separators=(",", ":"), allow_nan=False
-    )
-
-
-def _object(value: object, fields: set[str]) -> dict[str, object]:
-    if not isinstance(value, dict) or set(cast(dict[object, object], value)) != fields:
-        raise ValueError(f"消息对象字段必须为 {sorted(fields)}")
-    return cast(dict[str, object], value)
-
-
-def _part(value: object) -> Part:
-    if not isinstance(value, dict):
-        raise ValueError("消息 part 必须是对象")
-    raw = cast(dict[str, object], value)
-    if raw.get("kind") == "tool_call":
-        data = _object(raw, {"kind", "binding_id", "arguments"})
-        return ToolCall(
-            cast(str, data["binding_id"]), cast(Mapping[str, object], data["arguments"])
-        )
-    data = _object(raw, {"kind", "value"})
-    return ContentPart(cast(str, data["kind"]), data["value"])
-
-
-def _unique_fields(pairs: list[tuple[str, object]]) -> dict[str, object]:
-    result: dict[str, object] = {}
-    for key, value in pairs:
-        if key in result:
-            raise ValueError(f"消息 JSON 包含重复字段: {key}")
-        result[key] = value
-    return result
-
-
-def decode_body(payload: str) -> Body:
-    """在持久化边界拒绝未知结构；值域与不变量归消息构造函数。"""
-    raw: object = json.loads(payload, object_pairs_hook=_unique_fields)
-    if not isinstance(raw, dict):
-        raise ValueError("消息 body 必须是对象")
-    data = cast(dict[str, object], raw)
-    kind = data.get("kind")
-    if kind == "control":
-        row = _object(data, {"kind", "action", "through_seq", "reason"})
-        return Control(
-            cast(Literal["pause", "resume", "abandon", "failure"], row["action"]),
-            cast(int, row["through_seq"]),
-            cast(str | None, row["reason"]),
-        )
-    fields = {
-        "input": {"kind", "parts"},
-        "output": {"kind", "parts", "finish"},
-        "tool_result": {"kind", "parts", "call_ref", "outcome"},
-    }
-    if not isinstance(kind, str) or kind not in fields:
-        raise ValueError("消息 body kind 无效")
-    row = _object(data, fields[kind])
-    raw_parts = row["parts"]
-    if not isinstance(raw_parts, list):
-        raise ValueError("消息 parts 必须是数组")
-    parts = tuple(_part(part) for part in cast(list[object], raw_parts))
-    if kind == "input":
-        return Input(cast(tuple[ContentPart, ...], parts))
-    if kind == "output":
-        return Output(
-            parts,
-            cast(Literal["continue", "complete", "quiet"], row["finish"]),
-        )
-    call = _object(row["call_ref"], {"message_id", "part_index"})
-    result = ToolResult(
-        CallRef(cast(str, call["message_id"]), cast(int, call["part_index"])),
-        cast(Literal["success", "denied", "error", "interrupted"],
-             "error" if row["outcome"] == "unknown" else row["outcome"]),
-        cast(tuple[ContentPart, ...], parts),
-    )
-
-    if row["outcome"] == "unknown":
-        object.__setattr__(result, "_legacy_unknown", True)
-    return result
